backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
openpyxl==3.1.5
phonenumbers==9.0.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
